+++ sources/Windows/GoogleMusic/Views/PlaylistView.xaml.cs
@@ -143,7 +143,7 @@ namespace OutcoldSolutions.GoogleMusic.Views
 
         private void RatingOnValueChanged(object sender, ValueChangedEventArgs e)
         {
-            this.Presenter<PlaylistViewPresenter>().UpdateRating((Song)((Rating)sender).DataContext, e.NewValue);
+            this.Presenter<PlaylistViewPresenter>().UpdateRating((Song)((Rating)sender).DataContext, (byte)e.NewValue);
         }
     }
 }
